[tls] Answer empty client certificate if we don't have any
When a server requests a client certificate, but no suitable certificate
is available, the rfc stipulates that "the client SHOULD send a certificate
message containing no certificates".

Do this by just leaving `tls->certs` empty in
`tls_new_certificate_request()` instead of erroring out,
and only send a certificate verifymessage if we did find a certificate.

Signed-off-by: Laura Peeters <peeters.laura@ugent.be>